ファイル名変更，不要なファイルの削除．
