+++ README.md
@@ -12,46 +12,27 @@
 - インターネット環境
 - Ruby
 開発環境は ruby 2.1.5p273．version 2.0.0 以上を推奨．
-- Rubyライブラリ nokogiri
-HTMLをパースし，値を取り出すために使用する．
-- Rubyライブラリ anemone
-Webページを操作し，ページを移動する際に使用する．
 
 ## 使い方
 
 1. Ruby をインストールする
+
   (省略)
 
-2. Rubyライブラリをインストールする
-  (詳細は省略)
+2. 年度，所属コードの指定して実行する．
 
   ```
-  $ gem install anemone
-  $ gem install nokogiri
+  $ ruby FindSyllabus.rb
   ```
 
-3. 年度，学部の指定
   FindSyllabus.rb をテキストエディタで開き，
   年度と学部を指定する．
 
-  ```
-  # 年度
-  year = "2014"
-  
-  # 学部コード
-  #   シラバスの学部コード
-  #     教養教育: 58, 文学部: 05,教育学部: 07,法学部: 15,
-  #     理学部: 22,医学部: 42,薬学部: 44,工学部: 25
-  scd = "25"
-  ```
 
 4. 実行
 
   コマンドラインを開き，以下を実行する．
 
-  ```
-  $ ruby FindSyllabus.rb
-  ```
   
   出力結果をファイルに保存したい場合は，以下のように > を用いてリダイレクトする．
   作成されたファイルの文字コードはUTF-8になる．なお， filename.csv は
